+++ packages/e2e-test-utils/CHANGELOG.md
@@ -1,5 +1,9 @@
 ## Master
 
+### Enhancements
+
+- `visitAdminPage` will now throw an error (emit a test failure) when there are unexpected errors on hte page.
+
 ### New Features
 
 - Added `getPageError` function, returning a promise which resolves to an error message present in the page, if any exists.